conversions are not performed for the assignment operators

+++ main.c
@@ -4,8 +4,8 @@ main() {
 
    int  i = 17;
    char c = 'c'; /* ascii value is 99 */
-   int sum;
+   float sum;
 
    sum = i + c;
-   printf("Value of sum : %d\n", sum );
+   printf("Value of sum : %f\n", sum );
 }
